pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lambda_packer"
version='0.1.58'
description = "A tool to package Python AWS Lambda functions with zips, Docker containers, and layers."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

dependencies = [
    "Click",
    "PyYAML",
    "docker>=4.4.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-mock", "black", "twine", "bump2version"]

[project.scripts]
lambda-packer = "lambda_packer.cli:main"
